major refactor into separate files
Keeps the bloat down in the main file
